# frontend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Your API",
    "description": "Your API description",
    "termsOfService": "https://www.google.com/policies/terms/",
    "contact": {
      "email": "contact@yourdomain.com"
    },
    "license": {
      "name": "BSD License"
    },
    "version": "v1"
  },
  "host": "pos.goavto.uz",
  "schemes": ["http", "https"],
  "basePath": "/",
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {your_token}\""
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/logs/": {
      "get": {
        "operationId": "logs_list",
        "description": "",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "A search term.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "required": ["count", "results"],
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "next": {
                  "type": "string",
                  "format": "uri",
                  "x-nullable": true
                },
                "previous": {
                  "type": "string",
                  "format": "uri",
                  "x-nullable": true
                },
                "results": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UnifiedLog"
                  }
                }
              }
            }
          }
        },
        "tags": ["logs"]
      },
      "parameters": []
    },
    "/logs/{id}/": {
      "get": {
        "operationId": "logs_read",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/UnifiedLog"
            }
          }
        },
        "tags": ["logs"]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/product-batch/": {
      "get": {
        "operationId": "product-batch_list",
        "description": "",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "A search term.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "required": ["count", "results"],
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "next": {
                  "type": "string",
                  "format": "uri",
                  "x-nullable": true
                },
                "previous": {
                  "type": "string",
                  "format": "uri",
                  "x-nullable": true
                },
                "results": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ProductBatch"
                  }
                }
              }
            }
          }
        },
        "tags": ["product-batch"]
      },
      "post": {
        "operationId": "product-batch_create",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductBatch"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProductBatch"
            }
          }
        },
        "tags": ["product-batch"]
      },
      "parameters": []
    },
    "/product-batch/{id}/": {
      "get": {
        "operationId": "product-batch_read",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProductBatch"
            }
          }
        },
        "tags": ["product-batch"]
      },
      "patch": {
        "operationId": "product-batch_partial_update",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductBatch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProductBatch"
            }
          }
        },
        "tags": ["product-batch"]
      },
      "delete": {
        "operationId": "product-batch_delete",
        "description": "",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["product-batch"]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "A unique integer value identifying this product batch.",
          "required": true,
          "type": "integer"
        }
      ]
    },
...
    "/products/{sku}/": {
      "get": {
        "operationId": "products_read",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        },
        "tags": ["products"]
      },
      "patch": {
        "operationId": "products_partial_update",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        },
        "tags": ["products"]
      },
      "delete": {
        "operationId": "products_delete",
        "description": "",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["products"]
      },
      "parameters": [
        {
          "name": "sku",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/suppliers/{id}/": {
      "get": {
        "operationId": "suppliers_read",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        },
        "tags": ["suppliers"]
      },
      "patch": {
        "operationId": "suppliers_partial_update",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        },
        "tags": ["suppliers"]
      },
      "delete": {
        "operationId": "suppliers_delete",
        "description": "",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["suppliers"]
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "A unique integer value identifying this supplier.",
          "required": true,
          "type": "integer"
        }
      ]
    }
  },
  "definitions": {
...
    "Category": {
      "required": ["name"],
      "type": "object",
      "properties": {
        "name": {
          "title": "Name",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "id": {
          "title": "ID",
          "type": "integer",
          "readOnly": true
        },
        "created_at": {
          "title": "Created at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updated_at": {
          "title": "Updated at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "created_by": {
          "title": "Created by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        },
        "updated_by": {
          "title": "Updated by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        }
      }
    },
...
    "UnifiedLog": {
      "required": ["id", "type", "action", "created_at", "note"],
      "type": "object",
      "properties": {
        "id": {
          "title": "Id",
          "type": "integer"
        },
        "type": {
          "title": "Type",
          "type": "string",
          "minLength": 1
        },
        "action": {
          "title": "Action",
          "type": "string",
          "minLength": 1
        },
        "created_at": {
          "title": "Created at",
          "type": "string",
          "format": "date-time"
        },
        "created_by": {
          "title": "Created by",
          "type": "string",
          "readOnly": true
        },
        "note": {
          "title": "Note",
          "type": "string",
          "minLength": 1
        },
        "sku": {
          "title": "Sku",
          "type": "string",
          "readOnly": true
        },
        "author": {
          "title": "Author",
          "type": "string",
          "readOnly": true
        }
      }
    },
...
    "ProductBatch": {
      "required": ["quantity", "buy_price", "sell_price"],
      "type": "object",
      "properties": {
        "id": {
          "title": "ID",
          "type": "integer",
          "readOnly": true
        },
        "created_at": {
          "title": "Created at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "deleted": {
          "title": "Deleted",
          "type": "boolean"
        },
        "updated_at": {
          "title": "Updated at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "quantity": {
          "title": "Quantity",
          "type": "integer",
          "maximum": 2147483647,
          "minimum": 0
        },
        "buy_price": {
          "title": "Buy price",
          "type": "integer",
          "maximum": 9.223372036854776e18,
          "minimum": 0
        },
        "sell_price": {
          "title": "Sell price",
          "type": "integer",
          "maximum": 9.223372036854776e18,
          "minimum": 0
        },
        "exchange_rate": {
          "title": "Exchange rate",
          "type": "integer",
          "maximum": 9.223372036854776e18,
          "minimum": 0,
          "x-nullable": true
        },
        "created_by": {
          "title": "Created by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        },
        "updated_by": {
          "title": "Updated by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        },
        "product": {
          "title": "Product",
          "type": "integer"
        }
      }
    },
    "Supplier": {
      "required": ["full_name", "phone_number"],
      "type": "object",
      "properties": {
        "id": {
          "title": "ID",
          "type": "integer",
          "readOnly": true
        },
        "created_at": {
          "title": "Created at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "deleted": {
          "title": "Deleted",
          "type": "boolean"
        },
        "updated_at": {
          "title": "Updated at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "full_name": {
          "title": "Full name",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "phone_number": {
          "title": "Phone number",
          "type": "string",
          "maxLength": 20,
          "minLength": 1
        },
        "created_by": {
          "title": "Created by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        },
        "updated_by": {
          "title": "Updated by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        }
      }
    },
    "Product": {
      "required": [
        "product_batch",
        "supplier_id",
        "category_id",
        "name",
        "product_type"
      ],
      "type": "object",
      "properties": {
        "id": {
          "title": "ID",
          "type": "integer",
          "readOnly": true
        },
        "product_batches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductBatch"
          },
          "readOnly": true
        },
        "product_batch": {
          "$ref": "#/definitions/ProductBatch"
        },
        "total_quantity": {
          "title": "Total quantity",
          "type": "integer",
          "readOnly": true
        },
        "sell_price": {
          "title": "Sell price",
          "type": "integer",
          "readOnly": true
        },
        "supplier_id": {
          "title": "Supplier id",
          "type": "integer"
        },
        "supplier": {
          "$ref": "#/definitions/Supplier"
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "category_id": {
          "title": "Category id",
          "type": "integer"
        },
        "created_at": {
          "title": "Created at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "deleted": {
          "title": "Deleted",
          "type": "boolean"
        },
        "updated_at": {
          "title": "Updated at",
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "name": {
          "title": "Name",
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "description": {
          "title": "Description",
          "type": "string",
          "maxLength": 10000
        },
        "image": {
          "title": "Image",
          "type": "string",
          "readOnly": true,
          "x-nullable": true,
          "format": "uri"
        },
        "sku": {
          "title": "Sku",
          "type": "string",
          "readOnly": true,
          "minLength": 1
        },
        "product_type": {
          "title": "Product type",
          "type": "string",
          "enum": ["KG", "PIECE"]
        },
        "created_by": {
          "title": "Created by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        },
        "updated_by": {
          "title": "Updated by",
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "x-nullable": true
        }
      }
    },
...
  }
}